backend/requirements.txt
et_xmlfile==2.0.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
sqlalchemy ==2.0.25
psycopg2-binary ==2.9.8
python-dotenv==1.0.0
werkzeug==2.3.7
Flask==2.3.3
flask-cors==3.0.10
openai==0.27.8
requests==2.32.3
PyJWT==2.8.0
